Fix issue where directory watch would drop out. Remove printlns.
This fixes an issue where the file watcher would drop out. It turns out
cargo returns the path to the root module *file* rather than the src
*directory*, which makes sense in retrospect! By changing to the
directory the behaviour seems to fixed. notify-rs/notify#218

The debug printlns have been removed.

I did notice some other weird behaviour however. `cargo` will not
re-build and replace the target dylib while it is currently in use,
which also makes sense in retrospect. However, this is an issue if we
want to rebuild the target while using the previously built one. This
means that either ourselves or the users should make a tmp copy of the
generated target dylib before loading it.