Reload the cargo config and package on each file event
This is an attempt to work around an issue on macOS where consecutive
re-compiles of the package continually result in the same output, as if
the original src of the package from when `watch` is called is re-used
for every re-compile. By reloading all of cargo each time
`Watch::package` is called, we remove the possibility that this is
caused by holding onto the `cargo::core::Package` handle for the
lifetime of the `Watch`.